requirements.txt
click
gitpython
jinja2
pathlib2
sqlalchemy
